// frontend/src/app/shared/enum/actions.enum.ts
export enum UserPermissions {
  AddUser = 'AddUser',
  EditUser = 'EditUser',
  EditSelf = 'EditSelf',
  ShowUser = 'ShowUser',
  DeleteUser = 'DeleteUser',
}

export enum CustomerPermissions {
  AddCustomer = 'AddCustomer',
  EditCustomer = 'EditCustomer',
  ShowCustomer = 'ShowCustomer',
  DeleteCustomer = 'DeleteCustomer',
}

export enum ProjectPermissions {
  AddProject = 'AddProject',
  EditProject = 'EditProject',
  ShowProject = 'ShowProject',
  DeleteProject = 'DeleteProject',
  ShowProjectRole = 'ShowProjectRole',
  EditProjectRole = 'EditProjectRole',
}

export enum EditLeavePermissions {
  EditStatus = 'EditStatus',
  EditDescription = 'EditDescription',
  EditComment = 'EditComment',
  EditCalendar = 'EditCalendar',
  EditType = 'EditType',
  EditDate = 'EditDate',
}